have to test more

+++ init.c
@@ -6,7 +6,7 @@
 /*   By: wmonacho <wmonacho@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/23 14:03:10 by wmonacho          #+#    #+#             */
-/*   Updated: 2022/09/06 14:41:31 by wmonacho         ###   ########lyon.fr   */
+/*   Updated: 2022/09/07 12:58:35 by wmonacho         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	init_variable(t_param *param)
 	param->philo = NULL;
 }
 
-int	check_diff(unsigned long long last_eat)
+unsigned long long	check_diff(unsigned long long last_eat)
 {
 	return (gettime() - last_eat);
 }
